this is unneeded
git-svn-id: file:///Users/michaelkoepke/svn/semiologic@4196 6c9c31c4-cb3d-4424-af3d-876279169c7e

+++ plugins/sem-backend/trunk/schema/order-lines.procs.sql
@@ -9,7 +9,7 @@ DECLARE
 	p			record;
 	c			record;
 	t_ratio		numeric := 1;
-	cur_orders	float8;
+	cur_orders	float;
 	o_ratio		numeric := 1;
 BEGIN
 	IF	NEW.init_price IS NULL OR NEW.init_comm IS NULL OR
